refactor: move out DMS Merger
